move collision handling to separate script
