+++ ProtoGameUE/Assets/Scripts/EnemyLife.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b7890e60614fc5940bd199cbc3385e39
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
